feat: Add drawing effect
Added drawing effect using 'mouseenter' and not 'mouseover' because
the former only triggers the event once and we only need it once, while
the latter triggers it multiple times if moving over it and can be
maliciously abused to cause lag.

Also updated the background color through adding a special class so I
can exercise more using the classList property and avoid using inline
style.

Transferred the logic for populating the grid into a function, so I
can use it in a later step instead of rewriting old code and used it
to add event listeners instead of writing more unnecessary code.