feat: Add button for changing grid size
Added a button that can clear and change grid size to bigger or smaller
number. Numbers supported are 1 through 100. Any other number or text
defaults to 16x16 grid.(Warning: grid sizes above 30x30 have a delay
when being created because of massive numbers, especially when in the
high-end of the range)

Cleaned up the style a bit. Made it functional.