feat: Add random RGB painting effect
Added 'Extra Credit' exercise 1. Made it possible through mutual
exclusiveness of buttons. If one button is active another cannot be.

Also the populateGrid event listeners now always check if a button is
active or not. Instead of adding event listeners on every div for every
button, I used the same event listeners from the start.

Started the darkening exercise but didn't finish it.